// src/app/customer/model/customer.model.ts
export class CustomerModel{
    id: number;
    FullName: string;
    Gender: string;
    Profession: string;
    BirthDate: Date;
    Email: string;
    Website: string;
    AllowPromotionEmails: boolean;
    Address: string;
    Notes: string;
    City:string;
    TelephoneNumber: string[];
  }
  